requirements.txt
django==2.1.8
djangorestframework==3.9.1
gunicorn==19.9.0
